+++ package.json
@@ -6,7 +6,8 @@
     "dev": "open-cli http://localhost:3000 && next dev --hostname 0.0.0.0",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "clean": "rm -rf .next"
   },
   "dependencies": {
     "@emotion/react": "^11.11.0",
@@ -16,8 +17,11 @@
     "@fortawesome/free-regular-svg-icons": "^7.0.0",
     "@fortawesome/free-solid-svg-icons": "^7.0.0",
     "@fortawesome/react-fontawesome": "^0.2.3",
+    "@mdx-js/loader": "^3.1.0",
+    "@mdx-js/react": "^3.1.0",
     "@mui/icons-material": "^5.15.0",
     "@mui/material": "^5.15.0",
+    "@next/mdx": "^15.5.0",
     "framer-motion": "^12.23.12",
     "lenis": "^1.3.8",
     "lil-gui": "^0.20.0",
