+++ src/app/work/projects/[slug]/page.jsx
@@ -1,370 +1,14 @@
-'use client';
+import projects from "@/lib/data/projects";
+import ProjectModalClient from "./ProjectModalClient";
 
-import { useRouter } from 'next/navigation';
-import { IconButton, Box, Typography, Stack, Zoom, Fab } from '@mui/material';
-import { motion, AnimatePresence } from 'framer-motion';
-import ArrowBackIosIcon from '@mui/icons-material/ArrowBackIos';
-import CloseIcon from '@mui/icons-material/Close';
-import ArrowForwardIosIcon from '@mui/icons-material/ArrowForwardIos';
-import PROJECT_CARDS from '@/lib/data/projects';
-import { use, useEffect, useState, useRef, useCallback } from 'react';
-import React from 'react';
-import { useActiveCard } from '@/lib/hooks/useActiveCard';
-import { useScroll } from '@/context/ScrollContext';
-import Magnetic from '@/components/animations/Magnetic';
-import Image from "next/image";
-import Link from 'next/link';
+export async function generateStaticParams() {
+  return projects.map((p) => ({ slug: p.slug }));
+}
 
-export default function ProjectModalPage(props) {
-  const params = use(props.params);
-  const router = useRouter();
-  const project = PROJECT_CARDS.find((p) => p.slug === params.slug);
-  const { setActiveSlug } = useActiveCard();
-  const { setWrapper } = useScroll();
-  const modalRef = useRef(null);
-  const timerRef = useRef(null);
-  const [carouselIndex, setCarouselIndex] = useState(0);
-  const [lightboxIndex, setLightboxIndex] = useState(null);
-  const [isScrolled, setIsScrolled] = useState(false);
+export default function ProjectPage({ params }) {
+  const project = projects.find((p) => p.slug === params.slug);
 
-  const handleClose = () => {
-    router.push('/work/projects/');
-  };
+  if (!project) return <div>Project not found</div>;
 
-  useEffect(() => {
-    const scrollable = modalRef.current;
-    if (!scrollable) return;
-
-    const handleScroll = () => {
-      const scrollTop = scrollable.scrollTop;
-      setIsScrolled(scrollTop > 10);
-    };
-
-    scrollable.addEventListener('scroll', handleScroll);
-    handleScroll(); // Initial check
-
-    return () => {
-      scrollable.removeEventListener('scroll', handleScroll);
-    };
-  }, [modalRef]);
-
-  // Carousel controls
-  const showPrevCarousel = () => {
-    setCarouselIndex((i) => (i > 0 ? i - 1 : project.galleryImages.length - 1));
-    resetTimer();
-  };
-  const showNextCarousel = useCallback(() => {
-    setCarouselIndex((i) => (i < project.galleryImages.length - 1 ? i + 1 : 0));
-    resetTimer();
-  }, [project]);
-
-  // Lightbox controls
-  const closeLightbox = () => setLightboxIndex(null);
-  const showPrevLightbox = useCallback(
-    () => setLightboxIndex((i) => (i > 0 ? i - 1 : project.galleryImages.length - 1)),
-    [project]
-  );
-  const showNextLightbox = useCallback(
-    () => setLightboxIndex((i) => (i < project.galleryImages.length - 1 ? i + 1 : 0)),
-    [project]
-  );
-  
-  const resetTimer = () => {
-    if (timerRef.current) clearInterval(timerRef.current);
-    timerRef.current = setInterval(() => {
-      showNextCarousel();
-    }, 5000);
-  };
-
-  useEffect(() => {
-    if (project?.galleryImages?.length > 1) {
-      resetTimer();
-    }
-    return () => {
-      if (timerRef.current) clearInterval(timerRef.current);
-    };
-  }, [project, showNextCarousel]);
-
-
-  // Touch swipe handling
-  let touchStartX = 0;
-  const handleTouchStart = (e) => { touchStartX = e.touches[0].clientX; };
-  const handleTouchEnd = (e) => {
-    const deltaX = e.changedTouches[0].clientX - touchStartX;
-    if (deltaX > 50) showPrevLightbox();
-    else if (deltaX < -50) showNextLightbox();
-  };
-
-  useEffect(() => {
-    if (!project) return;
-    setActiveSlug(project.slug);
-    setWrapper(modalRef.current);
-    const onEsc = (e) => {
-      if (e.key === 'Escape') {
-        lightboxIndex !== null ? closeLightbox() : handleClose();
-      }
-    };
-    window.addEventListener('keydown', onEsc);
-    return () => {
-      window.removeEventListener('keydown', onEsc);
-      setActiveSlug(null);
-      setWrapper(null);
-    };
-  }, [project, lightboxIndex]);
-
-  
-
-
-  if (!project) return null;
-
-  // Combine paragraphs with midImages inserted evenly
-  const combinedContent = [];
-  const paragraphs = project.paragraphs || [project.info];
-  const midImages = project.midImages || [];
-  const insertEvery = Math.floor(paragraphs.length / (midImages.length + 1)) || paragraphs.length;
-  paragraphs.forEach((item, i) => {
-    if (typeof item === 'string') {
-      combinedContent.push({ type: 'text', content: item });
-    } else if (item.type === 'link') {
-      combinedContent.push({ type: 'link', ...item });
-    }
-    if (midImages.length && (i + 1) % insertEvery === 0 && midImages[(i + 1) / insertEvery - 1]) {
-      combinedContent.push({ type: 'image', ...midImages[(i + 1) / insertEvery - 1] });
-    }
-  });
-
-  const finalRenderedElements = [];
-  let i = 0;
-  while (i < combinedContent.length) {
-    const item = combinedContent[i];
-
-    if (item.type === 'text') {
-      const paragraphParts = [item.content];
-      let j = i + 1;
-
-      // Check for inline link pattern: text, link, text
-      if (j < combinedContent.length && combinedContent[j].type === 'link') {
-        const linkItem = combinedContent[j];
-        paragraphParts.push(
-          <Link key={`inline-link-${j}`} href={linkItem.href} style={{ color: 'var(--link-color)', textDecoration: 'underline' }}>
-            {linkItem.text}
-          </Link>
-        );
-        j++; // Move past the link
-
-        if (j < combinedContent.length && combinedContent[j].type === 'text') {
-          paragraphParts.push(combinedContent[j].content);
-          j++; // Move past the text after the link
-        }
-      }
-
-      finalRenderedElements.push(
-        <Typography key={`paragraph-${i}`} variant="body1" fontSize={'1.2rem'}>
-          {paragraphParts}
-        </Typography>
-      );
-      i = j; // Update outer loop index
-    } else if (item.type === 'image') {
-      finalRenderedElements.push(
-        <motion.img
-          key={`image-${i}`}
-          src={item.src}
-          alt={item.alt || ''}
-          style={{ width: '100%', borderRadius: 12, margin: '24px 0' }}
-          whileHover={{ scale: 1.02 }}
-        />
-      );
-      i++; // Move to next item
-    } else if (item.type === 'link') {
-      // This case should ideally be handled by the 'text' block looking ahead.
-      // If we reach a 'link' directly, it means it's not part of the expected pattern.
-      // For now, we'll just skip it or render it as a standalone link (which is not what we want).
-      // This indicates a problem with the combinedContent structure or the parsing logic.
-      i++;
-    }
-  }
-
-  const carouselVariants = {
-    enter: (direction) => ({
-      x: direction > 0 ? 1000 : -1000,
-      opacity: 0,
-    }),
-    center: {
-      zIndex: 1,
-      x: 0,
-      opacity: 1,
-    },
-    exit: (direction) => ({
-      zIndex: 0,
-      x: direction < 0 ? 1000 : -1000,
-      opacity: 0,
-    }),
-  };
-
-  return (
-      <motion.div
-        ref={modalRef}
-        layoutId={`card-${project.slug}`}
-        initial={{ scale: 0.9, opacity: 0 }}
-        animate={{ scale: 1, opacity: 1 }}
-        exit={{ scale: 0.9, opacity: 0 }}
-        transition={{ duration: 0.4 }}
-        style={{
-          position: 'fixed',
-          top: 0,
-          left: 0,
-          width: '100vw',
-          height: '100vh',
-          background: 'var(--background-color)',
-          padding: '40px',
-          overflowY: 'auto',
-          zIndex: 1300,
-          color: 'var(--primary-text)',
-        }}
-      >
-        <Stack direction="column" sx={{ justifyContent: 'center', alignItems: 'center' }}>
-          {/* Header Buttons */}
-          <Box sx={{ display: 'flex', justifyContent: 'start', mb: 3, width: '100%', height: '48px' }}>
-             <Image
-                  src={project.thumbnail}
-                  alt="Project Thumbnail"
-                  width={50}
-                  height={50}
-                  style={{ aspectRatio: '1 / 1', objectFit: 'fill' }}
-              />
-            <Box sx={{
-              position: 'fixed',
-              top: 40,
-              right: 60,
-              zIndex: 1301,
-            }}>
-              <Magnetic>
-                <IconButton 
-                  onClick={handleClose} 
-                  aria-label="close" 
-                  sx={{
-                    color: isScrolled ? 'var(--background-color)' : 'var(--primary-text)',
-                    backgroundColor: isScrolled ? 'var(--primary-text)' : 'transparent',
-                    boxShadow: isScrolled ? '3' : 'none',
-                    backdropFilter: isScrolled ? 'blur(10px)' : 'none',
-                    transition: 'all 0.2s ease-in-out',
-                    '&:hover': {
-                      backgroundColor: isScrolled ? 'var(--primary-text)' : 'transparent',
-                      color: isScrolled ? 'var(--secondary-text)' : 'var(--primary-text)',
-                      strokeWidth: '2'
-                    }
-                  }}
-                >
-                  <CloseIcon fontSize="large" />
-                </IconButton>
-              </Magnetic>
-            </Box>
-          </Box>
-
-          {/* Image Carousel */}
-          {project.galleryImages?.length > 0 && (
-            <Box sx={{ position: 'relative', width: {xs: '100%', md: '70%', lg: '50%'}, height: '500px', mb: 3, overflow: 'hidden', justifyContent: 'center', alignItems: 'center' }}>
-              <AnimatePresence initial={false}>
-              <motion.img
-                key={carouselIndex}
-                src={project.galleryImages[carouselIndex].src}
-                alt={project.galleryImages[carouselIndex].alt}
-                custom={1}
-                variants={carouselVariants}
-                initial="enter"
-                animate="center"
-                exit="exit"
-                transition={{
-                  x: { type: 'spring', stiffness: 100, damping: 30 },
-                  opacity: { duration: 0.2 },
-                }}
-                style={{
-                  position: 'absolute',
-                  top: 0,
-                  left: 0,
-                  width: '100%',
-                  height: '100%',
-                  objectFit: 'cover',
-                  borderRadius: 8,
-                  cursor: 'pointer',
-                }}
-                onClick={() => setLightboxIndex(carouselIndex)}
-              />
-            </AnimatePresence>
-              {project.galleryImages.length > 1 && (
-                <>
-                  <IconButton onClick={showPrevCarousel} sx={{ position: 'absolute', top: '50%', left: 16, color: 'white', transform: 'translateY(-50%)', zIndex: 2 }}>
-                    <ArrowBackIosIcon />
-                  </IconButton>
-                  <IconButton onClick={showNextCarousel} sx={{ position: 'absolute', top: '50%', right: 16, color: 'white', transform: 'translateY(-50%)', zIndex: 2 }}>
-                    <ArrowForwardIosIcon />
-                  </IconButton>
-                </>
-              )}
-            </Box>
-          )}
-
-          <motion.h2 layoutId={`card-title-${project.slug}`}>{project.title}</motion.h2>
-          <Typography variant="body1">
-            <strong>Technologies:</strong> {project.tech}
-          </Typography>
-
-          {/* Dynamic Content (Paragraphs + Mid Images) */}
-          <Stack spacing={2} sx={{ mt: 2 }}>
-            {finalRenderedElements}
-          </Stack>
-
-          {/* Fullscreen Lightbox */}
-          <AnimatePresence>
-            {lightboxIndex !== null && (
-              <motion.div
-                className="lightbox"
-                style={{
-                  position: 'fixed',
-                  top: 0,
-                  left: 0,
-                  width: '100vw',
-                  height: '100vh',
-                  background: 'rgba(0,0,0,0.9)',
-                  display: 'flex',
-                  flexDirection: 'column',
-                  justifyContent: 'center',
-                  alignItems: 'center',
-                  zIndex: 2000,
-                }}
-                initial={{ opacity: 0 }}
-                animate={{ opacity: 1 }}
-                exit={{ opacity: 0 }}
-                onClick={closeLightbox}
-                onTouchStart={handleTouchStart}
-                onTouchEnd={handleTouchEnd}
-              >
-                <img
-                  src={project.galleryImages[lightboxIndex].src}
-                  alt={project.galleryImages[lightboxIndex].alt}
-                  style={{ maxHeight: '80%', maxWidth: '90%', objectFit: 'contain' }}
-                />
-                {project.galleryImages[lightboxIndex].caption && (
-                  <Typography variant="body2" sx={{ color: '#fff', mt: 2 }}>
-                    {project.galleryImages[lightboxIndex].caption}
-                  </Typography>
-                )}
-                <IconButton
-                  onClick={(e) => { e.stopPropagation(); showPrevLightbox(); }}
-                  sx={{ position: 'absolute', left: 16, color: '#fff' }}
-                >‹</IconButton>
-                <IconButton
-                  onClick={(e) => { e.stopPropagation(); showNextLightbox(); }}
-                  sx={{ position: 'absolute', right: 16, color: '#fff' }}
-                >›</IconButton>
-                <IconButton
-                  onClick={(e) => { e.stopPropagation(); closeLightbox(); }}
-                  sx={{ position: 'absolute', top: 16, right: 16, color: '#fff' }}
-                >×</IconButton>
-              </motion.div>
-            )}
-          </AnimatePresence>
-        </Stack>
-      </motion.div>
-  );
+  return <ProjectModalClient project={project} />;
 }
