+++ src/lib/content/projects/eye-of-metamorphosis.mdx
@@ -1,0 +1,18 @@
+"use client"
+
+# Rocket Simulator
+
+A **real-time** WebGL-based rocket simulation using ThreeJS.
+
+![Rocket launch sequence](/images/rocket1.png)
+
+## Features
+- Real-time telemetry visualization
+- Interactive camera controls
+
+Check out the [GitHub repo](https://github.com/yourname/rocket-sim)  
+Or see the [live demo](/demo/rocket-sim).
+
+---
+
+<CustomButton href="https://rocket-sim.com">Launch Simulator</CustomButton>
